GL.TexImage2D呼び出しから不要なキャストを削除

+++ MiniMikuDanceMaui/PmxRenderer.cs
@@ -1055,7 +1055,7 @@ void main(){
         var handle = GCHandle.Alloc(bytes, GCHandleType.Pinned);
         try
         {
-            GL.TexImage2D(All.Texture2D, 0, (int)All.Rgba, width, height, 0, All.Rgba, All.UnsignedByte, handle.AddrOfPinnedObject());
+            GL.TexImage2D(All.Texture2D, 0, All.Rgba, width, height, 0, All.Rgba, All.UnsignedByte, handle.AddrOfPinnedObject());
             CheckGLError("GL.TexImage2D");
         }
         finally
